+++ forge/src/main/java/io/github/flemmli97/fateubw/forge/data/BlockStates.java
@@ -1,15 +1,12 @@
 package io.github.flemmli97.fateubw.forge.data;
 
 import io.github.flemmli97.fateubw.Fate;
-import io.github.flemmli97.fateubw.common.blocks.ChalkBlock;
 import io.github.flemmli97.fateubw.common.registry.ModBlocks;
 import net.minecraft.core.Direction;
 import net.minecraft.data.DataGenerator;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.Block;
 import net.minecraftforge.client.model.generators.BlockStateProvider;
-import net.minecraftforge.client.model.generators.ConfiguredModel;
-import net.minecraftforge.client.model.generators.ModelFile;
 import net.minecraftforge.common.data.ExistingFileHelper;
 
 public class BlockStates extends BlockStateProvider {
@@ -20,9 +17,11 @@ public class BlockStates extends BlockStateProvider {
 
     @Override
     protected void registerStatesAndModels() {
-        this.chalkBase();
-
-        this.chalk(ModBlocks.CHALK.get());
+        this.simpleBlock(ModBlocks.CHALK.get(), this.models().singleTexture("fateubw:block/" + ModBlocks.CHALK.getID().getPath(), this.mcLoc("block/thin_block"), "texture", new ResourceLocation(Fate.MODID, "blocks/" + ModBlocks.CHALK.getID().getPath()))
+                .texture("particle", new ResourceLocation(Fate.MODID, "blocks/" + ModBlocks.CHALK.getID().getPath())).element().from(0, 0, 0).to(16, 0.5f, 16)
+                .face(Direction.UP).uvs(0, 0, 16, 16).texture("#texture").end()
+                .face(Direction.DOWN).uvs(0, 16, 16, 0).cullface(Direction.DOWN).texture("#texture").end()
+                .end());
         this.simpleBlock(ModBlocks.ARTIFACT_ORE.get(), this.models().singleTexture(ModBlocks.ARTIFACT_ORE.getID().getPath(), new ResourceLocation(Fate.MODID, "block/ore"), "ore", new ResourceLocation(Fate.MODID, "blocks/charm_ore")));
         this.simpleBlock(ModBlocks.GEM_ORE.get(), this.models().singleTexture(ModBlocks.GEM_ORE.getID().getPath(), new ResourceLocation(Fate.MODID, "block/ore"), "ore", new ResourceLocation(Fate.MODID, "blocks/crystal_ore")));
         this.simpleBlock(ModBlocks.DEEP_SLATE_ARTIFACT_ORE.get(), this.models().getBuilder(ModBlocks.DEEP_SLATE_ARTIFACT_ORE.getID().getPath())
@@ -40,24 +39,5 @@ public class BlockStates extends BlockStateProvider {
     public ResourceLocation blockTexture(Block block) {
         ResourceLocation name = block.getRegistryName();
         return new ResourceLocation(name.getNamespace(), "blocks" + "/" + name.getPath());
-    }
-
-    private void chalk(ChalkBlock block) {
-        this.getVariantBuilder(block)
-                .forAllStates(state -> {
-                    String model = block.getRegistryName().getPath() + "_" + state.getValue(ChalkBlock.POSITION).getID();
-                    return ConfiguredModel.builder()
-                            .modelFile(this.models().singleTexture(model, new ResourceLocation(Fate.MODID, "block/chalk_base"), new ResourceLocation(block.getRegistryName().getNamespace(), "blocks/" + model)))
-                            .rotationY((int) (state.getValue(ChalkBlock.FACING).toYRot() + 180) % 360)
-                            .build();
-                });
-    }
-
-    private ModelFile chalkBase() {
-        return this.models().singleTexture("fateubw:block/chalk_base", this.mcLoc("block/thin_block"), "particle", new ResourceLocation(Fate.MODID, "blocks/chalk_line_none"))
-                .element().from(0, 0, 0).to(16, 0.5f, 16)
-                .face(Direction.UP).uvs(0, 0, 16, 16).texture("#texture").end()
-                .face(Direction.DOWN).uvs(0, 16, 16, 0).cullface(Direction.DOWN).texture("#texture").end()
-                .end();
     }
 }