+++ forge/src/main/java/io/github/flemmli97/fateubw/forge/data/ItemModels.java
@@ -128,10 +128,9 @@ public class ItemModels extends ItemModelProvider {
             } else if (reg == ModItems.CRYSTAL_RED || reg == ModItems.CRYSTAL_YELLOW || reg == ModItems.CRYSTAL_BLACK || reg == ModItems.CRYSTAL_BLUE || reg == ModItems.CRYSTAL_GREEN) {
                 this.withExistingParent(reg.getID().getPath(), new ResourceLocation(Fate.MODID, "item/gem_shard"))
                         .texture("layer0", this.itemTexture(reg.getID()));
-            } else if(reg == ModItems.ANIMATION_DEBUG) {
+            } else if (reg == ModItems.ANIMATION_DEBUG) {
                 this.singleTexture(reg.getID().getPath(), this.mcLoc("item/generated"), "layer0", new ResourceLocation("item/stick"));
-            }
-            else {
+            } else {
                 this.withExistingParent(reg.getID().getPath(), ModelLocationUtils.decorateItemModelLocation("generated"))
                         .texture("layer0", this.itemTexture(reg.getID()));
             }